+++ R/ggplot2.r
@@ -90,3 +90,24 @@ warning_wrap <- function(...) {
   warning(paste0(wrapped, collapse = "\n"), call. = FALSE)
 }
 
+is.formula <- function(x) inherits(x, "formula")
+
+# From ggplot2/R/axis-secondary.R
+is.sec_axis <- function(x) {
+  inherits(x, "AxisSecondary")
+}
+
+
+# From ggplot2/R/theme.r
+# renders a theme as complete
+#' @importFrom plyr defaults
+plot_theme <- function (x, default = theme_get()) {
+  theme <- x$theme
+  if (is_theme_complete(theme)) {
+    theme
+  }
+  else {
+    plyr::defaults(theme, default)
+  }
+}
+is_theme_complete <- function(x) isTRUE(attr(x, "complete"))